backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stock-reason-backend"
version = "0.1.0"
description = "股票事件追踪系统后端"
requires-python = ">=3.10,<=3.12"
dependencies = [
    "fastapi==0.103.1",
    "uvicorn==0.23.2",
    "sqlalchemy==2.0.20",
    "pydantic==2.3.0",
    "pydantic-settings>=2.0.0",
    "alembic==1.12.0",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "httpx==0.25.0",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "python-multipart==0.0.6",
    "redis==5.0.0",
    "tenacity==8.2.3",
    "asyncio==3.4.3",
    "aioredis==2.0.1",
    "python-dotenv==1.0.0",
    "pytest==7.4.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "ruff",
]

[tool.pytest]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"